+++ src/tools/__init__.py
@@ -10,7 +10,6 @@ from .pku3b_announcements import PKU3bAnnouncementsTool
 from .pku3b_assignments import PKU3bAssignmentsTool
 from .pku3b_coursetable import PKU3bCourseTableTool
 from .plib_materials import PLibMaterialsTool
-from .reminder import ReminderTool
 from .treehole_updates import TreeholeAuthService, TreeholeTool, TreeholeUpdatesTool
 
 __all__ = [
@@ -25,7 +24,6 @@ __all__ = [
     "PKU3bAnnouncementsTool",
     "PKU3bAssignmentsTool",
     "PKU3bCourseTableTool",
-    "ReminderTool",
     "Tool",
     "ToolRegistry",
     "ToolResult",